Python config validation mega-PR (#1996)
* start copying over work stored in issue comments

* Fix naming

* Snake case all over in python

* Move the rest too

* Clang tidy and ruff fix after the changes

* build out module of python config

- add parameter_set testing to ctest
- separate different logical units into their own files for organization
- add __init__ to load them for downstream users
- update cmake to support underscore files
- revert ldmxcfg.py back to trunk to enable legacy config to still function

* some patches noticed after inspecting online

* Clang format

* Ruff lint

* Fix the other configs, and logging names

* patch up how we deduce python modules

* typos caught while trying to get tests started

* WIP: will not function but caching work

* Fix basic test and many other CIs

* clang format

* Fix total_events too

* Clang tidy

* Apply clang-format and clang-tidy

* move instance_name definition to decorator

* replace old file with updated module

* start rename from camelCase to snake_case

* current cmake does not support nested py modules

* finish conversion to snake_case in Framework

- need to hack to make sure all dataclass fields are present in __dict__
- remove test on bad param in ConfigurePythonTest, new python-only test
  checks for bad param checking
- update parameter_set to function with a 1D-list required parameter

* refactor python test to make sure the __dict__ member is correct

this is the member variable that we inspect on the C++ side and so we
need it to be a dictionary of the parameters we want to pass into C++
Parameters

* make sure we can change required parameters

* avoid need for field for other mutable default objects

* loop over annotations, better list handling and error message

* convert E/H cal Geometry conditions as examples

* found old parameter now with checking!

* focus on enabling pass through of dataclasses.field

- helps be more transparent with how the annotations are being
  interpreted
- do not impose defaults, defaults are the same as those for
  dataclasses.field
- user can specify init=False if they want

* make sure required parameters are required

* WIP: re-introduce required base class for certain types of parameter
sets

this is helpful for type-deduction and type checking, I wasn't able to
find a easy/robust way of doing this entirely within the decorator and
so I'm still going to require the base class for some parameter sets

most parameter sets that aren't the common Processor/COP types don't
need to have any base class and can just use the decorator to still get
the existence/type checking benefits

* apply ruff formatting to updated modules

* WIP: good progress translating, new wrinkle

want to be able to have derived classes have required arguments

* avoid req after opt by not having base-class members

have them added by decorator

new wrinkle, dynamic init=False member not present for full_post_init

* able to get passed config load for unbiased simulation

- allow inheritance in list entry type checking
- allow None when copying dataclass_fields to dict
- define last_process earlier and dedup conditions_object registry
- fixup post_init of SDs

* fixed a few more things, now EcalGeo not calling base post_init

* parent class functions are over-ridded by child functions

this makes sense I just forgot, dropping post_inits from parent and
putting them in decoratator where it can be _combined_ with a
user-defined post_init that is optional

* Framework, Run_SimCore_basic, PythonConfigParameterSet tests pass

* bring conditions test up to date with new infra

* remove old printout

* apply ruff format and ruff check --fix

* update test to pass ruff check

* get ruff linting to pass on Framework/python

* auto format and lint by ruff

cd SimCore
denv ruff format
denv ruff check --fix --unsafe-fixes

* part way through manual ruff updates, had an idea

* remove additional checking on sum-to-one

I couldn't find a way to make it not clunky without changing everything
else as well, so just removing the check

* finish ruff check ignoring N (naming) checks

* fixup histogram name

* update some doc strings in parameter_set

* add tests for Standalone processors and make_processor

* add doc string detail to complicated python

* convert biasing utilities

* fixup biasing_utility function after test run

* convert filters as well

* convert Biasing filters and do ruff format/lint

* update instance_name, patch biasing operator decorator

* start conversion of Hcal module

* first draft at converting digi.py

* draft hcal translation and ruff format

* convert hcal_ana module

* finish first draft and ruff format everything

* mostly auto and some manual ruff check lints

* now passing Hcal basic tests

* translate Ecal to new style, Ecal basic test pass

* format and some basic lints

* more ruff format and lints

* move PrimaryGeneratorAction into SimCore out of G4User

explain in documentation to avoid making same error in the future

* G4User uses Event objects so it needs to link to Event

* remove leftover debug print statement

* Migrating the tracking modules

* Migrating the ts modules

* Migrating the Recon modules

* Migrating the Recon modules Part 2

* Ecal config CI runs fine at this point

* Python lint

* Migrate Packing

* Migrate Tools

* Migrate the Trigger modules

* Apply clang-format and clang-tidy

* Make the rest of the CIs pass too

* One more attempt to fix the rest of the CIs

* Attempt No3

* Attempt 4

* Fix sim_hit_pass_name, and float double conversion

* Fix some mistakes brought in the merge

* Add protection for tracking DQM if no truth track exists

* Fixes for HCAL and GENIE CI

* Apply clang-format and clang-tidy

* Fix truth track names

* use class base name (drop namespace) for default instance name

* add test to confirm that we can change instance_name (we can)

* fixup how default instance_name is set and accessed

also move shared post_init into decorator so it isn't silently
overwritten by downstream processors

* share post_init with make_processor as well

---------

+++ Trigger/Algo/src/Trigger/TrigMipReco.cxx
@@ -14,19 +14,19 @@ void TrigMipReco::configure(framework::config::Parameters& ps) {
 
   max_layer_ = ps.get<int>("max_layer", 32);
   // mm
-  search_radius_ = ps.get<float>("search_radius", 50.0f);
+  search_radius_ = ps.get<double>("search_radius", 50.0);
   min_track_length_ = ps.get<int>("min_track_length", 5);
   // MeV; Change as needed
-  isolation_e_cut_ = ps.get<float>("isolation_e_cut", 180.0f);
-  hole_fraction_max_ = ps.get<float>("hole_fraction_max", 0.2f);
+  isolation_e_cut_ = ps.get<double>("isolation_e_cut", 180.0);
+  hole_fraction_max_ = ps.get<double>("hole_fraction_max", 0.2);
 
   if (calorimeter_type_is_hcal_) {
     // MIP peak is 10-11 MeV
-    hcal_min_energy_ = ps.get<float>("hcal_min_energy", 8.0f);
+    hcal_min_energy_ = ps.get<double>("hcal_min_energy", 8.0);
   } else {  // ECAL
     // MIP peak around 17 MeV
-    ecal_min_energy_ = ps.get<float>("ecal_min_energy", 3.0f);
-    ecal_max_energy_ = ps.get<float>("ecal_max_energy", 26.0f);
+    ecal_min_energy_ = ps.get<double>("ecal_min_energy", 3.0);
+    ecal_max_energy_ = ps.get<double>("ecal_max_energy", 26.0);
   }
 }
